pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "screenase"
version = "0.1.0"
description = "DoE tool for in-vitro transcription reaction optimization"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Ethan Arnold", email = "emailtoethan@gmail.com" }]
keywords = ["doe", "design-of-experiments", "ivt", "bioinformatics", "benchling"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "jinja2>=3.1",
    "pydantic>=2.6",
    "pyyaml>=6",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
    "types-PyYAML",
]
ui = [
    "streamlit>=1.32",
]
benchling = [
    "benchling-sdk>=1.15",
]

[project.scripts]
screenase = "screenase.cli:main"

[project.urls]
Homepage = "https://github.com/ethanarnold/screenase"
Repository = "https://github.com/ethanarnold/screenase"

[tool.hatch.build.targets.wheel]
packages = ["src/screenase"]

[tool.hatch.build.targets.sdist]
include = ["src/screenase", "README.md", "LICENSE", "pyproject.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
